(Refactor) import statements for stores
Modified the import statements across multiple components and utilities for better maintainability. The goal is to have a neater code base where all the necessary stores are imported directly from '../../stores'. Also, correct the wording from 'nevermind' to 'never mind' in change-email.svelte.
