+++ app/components/users/if-user.svelte
@@ -3,7 +3,7 @@
 <!-- Is signed in  -->
 
 <script lang="ts">
-  import { user } from '../../stores/user';
+  import { user } from '../../stores';
 </script>
 
 {#if $user}
